Triangulation Program
Program that generates a triangulation from a set of input data. Input
data relates to individual points (vertices) that comprise a 2-D space
comprised from individual triangles. The program is capable of file
input/output and queries of the form; which triangle contains a given
point. The project is intended as a base for mathematical simulations
over a 2-D space.